Taking carbon cycle functionality out of concentrations_emissions_handler and into separate module carbon_cycle_mod
